requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
groq
requests
geopy
chromadb
sentence-transformers
pydantic
